Moved Driver Metrics to separate class
