+++ src/main/scala/com/qubole/sparklens/analyzer/SimpleAppAnalyzer.scala
@@ -34,6 +34,7 @@ class SimpleAppAnalyzer extends  AppAnalyzer {
       "task-level granularity and aggregated across the app (all tasks, stages, and jobs).\n")
     ac.appMetrics.print("Application Metrics", out)
     out.println("\n")
+    ac.driverMetrics.print("Driver Metrics", out)
     out.toString()
   }
 }
